ai_video_pipeline/tools/elevenlabs.py
"""
ElevenLabs API wrapper for voice synthesis.

This module provides a wrapper around the ElevenLabs API for text-to-speech
functionality used by the Voice-Synthesis Agent.
"""

from typing import Dict, Any, Optional, BinaryIO
import os
import json
import requests
from pathlib import Path

from ai_video_pipeline.config.settings import settings


class ElevenLabsAPI:
    """
    Wrapper for the ElevenLabs text-to-speech API.
    
    This class provides methods to interact with the ElevenLabs API
    for generating realistic voice audio from text.
    """
    
    def __init__(self, api_key: Optional[str] = None):
        """
        Initialize the ElevenLabs API wrapper.
        
        Args:
            api_key: Optional API key (defaults to environment variable)
        """
        self.api_key = api_key or settings.elevenlabs_api_key
        self.base_url = "https://api.elevenlabs.io/v1"
        self.headers = {
            "Accept": "audio/mpeg",
            "Content-Type": "application/json",
            "xi-api-key": self.api_key
        }
    
    def list_voices(self) -> Dict[str, Any]:
        """
        List available voices.
        
        Returns:
            Dict[str, Any]: List of available voices
        """
        # This is a stub implementation that will be expanded in future sprints
        # In a real implementation, we would make an API call to ElevenLabs
        
        return {
            "voices": [
                {
                    "voice_id": "default",
                    "name": "Default Voice",
                    "preview_url": ""
                }
            ]
        }
    
    def synthesize_speech(
        self, 
        text: str, 
        voice_id: str = "default",
        output_path: Optional[Path] = None
    ) -> Path:
        """
        Synthesize speech from text.
        
        Args:
            text: Text to synthesize
            voice_id: ID of the voice to use
            output_path: Optional path to save the audio file
            
        Returns:
            Path: Path to the generated audio file
        """
        # This is a stub implementation that will be expanded in future sprints
        # In a real implementation, we would make an API call to ElevenLabs
        
        # If no output path is provided, create one in the assets directory
        if output_path is None:
            os.makedirs(settings.audio_dir, exist_ok=True)
            output_path = settings.audio_dir / f"{voice_id}_output.wav"
        
        # Create a placeholder file
        with open(output_path, "w") as f:
            f.write(f"# This is a placeholder for the generated audio file\n")
            f.write(f"# Text: {text}\n")
            f.write(f"# Voice ID: {voice_id}\n")
        
        return output_path
